[VC] Split Script.cs
