+++ server/GSCode.Parser/Script.cs
@@ -25,7 +25,7 @@ namespace GSCode.Parser;
 
 using SymbolKindSA = GSCode.Parser.SA.SymbolKind;
 
-public class Script(DocumentUri ScriptUri, string languageId, ISymbolLocationProvider? globalSymbolProvider = null, ScriptMode mode = ScriptMode.Editor)
+public partial class Script(DocumentUri ScriptUri, string languageId, ISymbolLocationProvider? globalSymbolProvider = null, ScriptMode mode = ScriptMode.Editor)
 {
     public bool Failed { get; private set; } = false;
     public bool Parsed { get; private set; } = false;
@@ -230,400 +230,6 @@ public class Script(DocumentUri ScriptUri, string languageId, ISymbolLocationPro
         return Task.CompletedTask;
     }
 
-    // --- Index-based helpers ---
-
-    /// <summary>
-    /// Index-based: checks if identifier token is preceded by &amp; (address-of operator).
-    /// </summary>
-    private bool IsAddressOfIdentifierByIndex(int tokenIndex)
-    {
-        var tokens = Sense.Tokens;
-        int leftMostIdx = tokenIndex;
-        // Check for ns::name pattern
-        int prevIdx = tokens.PrevNonTriviaIndex(leftMostIdx);
-        if (prevIdx >= 0 && tokens.GetAt(prevIdx)!.Type == TokenType.ScopeResolution)
-        {
-            int nsIdx = tokens.PrevNonTriviaIndex(prevIdx);
-            if (nsIdx >= 0 && tokens.GetAt(nsIdx)!.Type == TokenType.Identifier)
-            {
-                leftMostIdx = nsIdx;
-            }
-        }
-        int beforeIdx = tokens.PrevNonTriviaIndex(leftMostIdx);
-        return beforeIdx >= 0 && tokens.GetAt(beforeIdx)!.Type == TokenType.BitAnd;
-    }
-
-    /// <summary>
-    /// Index-based: parse namespace::name or just name from a token at the given index.
-    /// </summary>
-    private (string? qualifier, string name) ParseNamespaceQualifiedIdentifierByIndex(int tokenIndex)
-    {
-        var tokens = Sense.Tokens;
-        Token token = tokens.GetAt(tokenIndex)!;
-        int prevIdx = tokens.PrevNonTriviaIndex(tokenIndex);
-        if (prevIdx >= 0)
-        {
-            Token prev = tokens.GetAt(prevIdx)!;
-            if (prev.Type == TokenType.ScopeResolution)
-            {
-                int nsIdx = tokens.PrevNonTriviaIndex(prevIdx);
-                if (nsIdx >= 0)
-                {
-                    Token nsToken = tokens.GetAt(nsIdx)!;
-                    if (nsToken.Type == TokenType.Identifier)
-                    {
-                        return (nsToken.Lexeme, token.Lexeme);
-                    }
-                }
-            }
-        }
-        return (null, token.Lexeme);
-    }
-
-    /// <summary>
-    /// Index-based: checks if token is a namespace qualifier (followed by :: and identifier),
-    /// and returns the function token index after the ::.
-    /// </summary>
-    private bool TryGetQualifiedFunctionTokenByIndex(int tokenIndex, out int functionTokenIndex)
-    {
-        var tokens = Sense.Tokens;
-        functionTokenIndex = -1;
-
-        int nextIdx = tokens.NextNonTriviaIndex(tokenIndex);
-        if (nextIdx < 0 || tokens.GetAt(nextIdx)!.Type != TokenType.ScopeResolution)
-            return false;
-
-        int funcIdx = tokens.NextNonTriviaIndex(nextIdx);
-        if (funcIdx < 0 || tokens.GetAt(funcIdx)!.Type != TokenType.Identifier)
-            return false;
-
-        functionTokenIndex = funcIdx;
-        return true;
-    }
-
-    private bool IsOnUsingLineByIndex(int tokenIndex, out string? usingPath, out Range? usingRange)
-    {
-        var tokens = Sense.Tokens;
-        usingPath = null;
-        usingRange = null;
-
-        Token token = tokens.GetAt(tokenIndex)!;
-        int line = token.Range.Start.Line;
-
-        // Walk to start of line
-        int cursorIdx = tokenIndex;
-        while (cursorIdx > 0)
-        {
-            Token prev = tokens.GetAt(cursorIdx - 1)!;
-            if (prev.Range.End.Line != line) break;
-            cursorIdx--;
-        }
-
-        // Find #using on this line
-        int usingIdx = -1;
-        for (int i = cursorIdx; i < tokens.Count; i++)
-        {
-            Token t = tokens.GetAt(i)!;
-            if (t.Range.Start.Line != line) break;
-            if (t.Lexeme == "#using") { usingIdx = i; break; }
-        }
-        if (usingIdx < 0) return false;
-
-        // Collect tokens after #using up to ; or EOL
-        int startIdx = tokens.NextNonWhitespaceIndex(usingIdx);
-        if (startIdx < 0 || tokens.GetAt(startIdx)!.Range.Start.Line != line) return false;
-
-        Token startToken = tokens.GetAt(startIdx)!;
-        Token endToken = startToken;
-        for (int i = startIdx; i < tokens.Count; i++)
-        {
-            Token t = tokens.GetAt(i)!;
-            if (t.Range.Start.Line != line) break;
-            if (t.Type == TokenType.Semicolon || t.Type == TokenType.LineBreak) break;
-            endToken = t;
-        }
-
-        var sb = new StringBuilder();
-        for (int i = startIdx; ; i++)
-        {
-            Token t = tokens.GetAt(i)!;
-            if (t.Range.Start.Line != line) break;
-            if (t.Type == TokenType.Semicolon || t.Type == TokenType.LineBreak) break;
-            if (!t.IsWhitespacey()) sb.Append(t.Lexeme);
-            if (ReferenceEquals(t, endToken)) break;
-        }
-
-        usingPath = sb.ToString();
-        usingRange = RangeHelper.From(startToken.Range.Start, endToken.Range.End);
-        return true;
-    }
-
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static bool IsFunctionPointerCallIdentifier(Token identifier)
-    {
-        // Pattern: [[ identifier ]]( ... )
-        // Check immediate surrounding tokens ignoring trivia
-        Token? prev1 = identifier.PreviousNonTrivia();
-        if (prev1?.Type != TokenType.OpenBracket) return false;
-        Token? prev2 = prev1.PreviousNonTrivia();
-        if (prev2?.Type != TokenType.OpenBracket) return false;
-        Token? next1 = identifier.NextNonTrivia();
-        if (next1?.Type != TokenType.CloseBracket) return false;
-        Token? next2 = next1.NextNonTrivia();
-        if (next2?.Type != TokenType.CloseBracket) return false;
-        Token? next3 = next2.NextNonTrivia();
-        if (next3?.Type != TokenType.OpenParen) return false;
-        return true;
-    }
-
-    private void BuildReferenceIndex()
-    {
-        _references.Clear();
-        var api = TryGetApi();
-        var tokens = Sense.Tokens;
-        for (int i = 0; i < tokens.Count; i++)
-        {
-            Token token = tokens.GetAt(i)!;
-            if (token.Type != TokenType.Identifier) continue;
-
-            // Recognize definition identifiers
-            var senseDef = Sense.GetSenseDefinition(token);
-            if (senseDef is ScrFunctionSymbol)
-            {
-                var defNamespace = GetEffectiveNamespace();
-                AddRef(new SymbolKey(SymbolKindSA.Function, defNamespace, token.Lexeme), token.Range);
-                continue;
-            }
-            if (senseDef is ScrClassSymbol)
-            {
-                var defNamespace = GetEffectiveNamespace();
-                AddRef(new SymbolKey(SymbolKindSA.Class, defNamespace, token.Lexeme), token.Range);
-                continue;
-            }
-
-            // Recognize call-site or qualified references, or address-of '&name' / '&ns::name'
-            int nextIdx = tokens.NextNonWhitespaceIndex(i);
-            bool looksLikeCall = nextIdx >= 0 && tokens.GetAt(nextIdx)!.Type == TokenType.OpenParen;
-            int prevIdx = tokens.PrevNonTriviaIndex(i);
-            bool isQualified = prevIdx >= 0 && tokens.GetAt(prevIdx)!.Type == TokenType.ScopeResolution
-                && tokens.PrevNonTriviaIndex(prevIdx) is int nsPrevIdx && nsPrevIdx >= 0
-                && tokens.GetAt(nsPrevIdx)!.Type == TokenType.Identifier;
-            bool isAddressOf = IsAddressOfIdentifierByIndex(i);
-            if (!looksLikeCall && !isQualified && !isAddressOf) continue;
-
-            var (qual, name) = ParseNamespaceQualifiedIdentifierByIndex(i);
-
-            // Skip builtin
-            if (api is not null)
-            {
-                try { if (api.GetApiFunction(name) is not null) continue; } catch { }
-            }
-
-            // Resolve to a namespace
-            string resolvedNamespace = qual ?? GetEffectiveNamespace();
-            // Index as function reference for now (method support can be added later)
-            AddRef(new SymbolKey(SymbolKindSA.Function, resolvedNamespace, name), token.Range);
-        }
-
-        void AddRef(SymbolKey key, Range range)
-        {
-            if (!_references.TryGetValue(key, out var list))
-            {
-                list = new List<Range>();
-                _references[key] = list;
-            }
-            list.Add(range);
-        }
-    }
-
-    /// <summary>
-    /// Precomputes function scope info from the AST before it's disposed.
-    /// </summary>
-    private void PrecomputeFunctionScopes()
-    {
-        if (RootNode is null) return;
-        _functionScopes = new List<FunctionScopeInfo>();
-        foreach (var fn in EnumerateFunctions(RootNode))
-        {
-            string? name = fn.Name?.Lexeme;
-            Range bodyRange = GetStmtListRange(fn.Body);
-            var parameters = new List<(string Name, Range Range)>();
-            foreach (var p in fn.Parameters.Parameters)
-            {
-                if (p.Name is not null)
-                {
-                    parameters.Add((p.Name.Lexeme, p.Name.Range));
-                }
-            }
-            _functionScopes.Add(new FunctionScopeInfo(name, bodyRange, parameters));
-        }
-    }
-
-    public async Task<string?> GetEnclosingFunctionScopeIdAsync(Position position, CancellationToken cancellationToken = default)
-    {
-        await WaitUntilParsedAsync(cancellationToken);
-
-        // Use precomputed scopes if available (post-analysis, AST disposed)
-        if (_functionScopes is not null)
-        {
-            foreach (var scope in _functionScopes)
-            {
-                if (scope.FunctionName is null) continue;
-                if (IsPositionInsideRange(position, scope.BodyRange))
-                {
-                    return $"{GetEffectiveNamespace()}::{scope.FunctionName}";
-                }
-            }
-            return null;
-        }
-
-        // Fallback: use AST directly (pre-analysis)
-        if (RootNode is null) return null;
-        foreach (var fn in EnumerateFunctions(RootNode))
-        {
-            if (fn.Name is not Token nameTok) continue;
-            var bodyRange = GetStmtListRange(fn.Body);
-            if (IsPositionInsideRange(position, bodyRange))
-            {
-                return $"{GetEffectiveNamespace()}::{nameTok.Lexeme}";
-            }
-        }
-        return null;
-    }
-
-    private static Range GetStmtListRange(StmtListNode body)
-    {
-        if (body.Statements.Count == 0)
-        {
-            return RangeHelper.From(0, 0, 0, 0);
-        }
-        Range? start = null;
-        Range? end = null;
-        foreach (var st in body.Statements)
-        {
-            if (TryGetRange(st, out var r))
-            {
-                if (start is null) start = r;
-                end = r;
-            }
-        }
-        if (start is null || end is null)
-        {
-            return RangeHelper.From(0, 0, 0, 0);
-        }
-        return RangeHelper.From(start!.Start, end!.End);
-    }
-
-    private static bool IsPositionInsideRange(Position pos, Range range)
-    {
-        int cmpStart = ComparePosition(pos, range.Start);
-        int cmpEnd = ComparePosition(range.End, pos);
-        return cmpStart >= 0 && cmpEnd >= 0;
-    }
-
-    private static int ComparePosition(Position a, Position b)
-    {
-        if (a.Line != b.Line) return a.Line.CompareTo(b.Line);
-        return a.Character.CompareTo(b.Character);
-    }
-
-    private static bool TryGetRange(AstNode node, out Range range)
-    {
-        // Fast paths for nodes that carry a Range
-        switch (node)
-        {
-            case ExprNode e:
-                range = e.Range; return true;
-            case ControlFlowActionNode cfan:
-                range = cfan.Range; return true;
-            case ConstStmtNode cst:
-                range = cst.Range; return true;
-            case ExprStmtNode es when es.Expr is not null:
-                range = es.Expr.Range; return true;
-            case ReturnStmtNode rt when rt.Value is not null:
-                range = rt.Value.Range; return true;
-            case ArgsListNode al:
-                range = al.Range; return true;
-        }
-
-        // Composite statements: compute union of child ranges
-        bool ok = false;
-        Range start = default, end = default;
-        void Acc(Range r)
-        {
-            if (!ok) { start = r; end = r; ok = true; }
-            else { start = RangeHelper.From(start.Start, r.Start); end = RangeHelper.From(end.Start, r.End); }
-        }
-
-        switch (node)
-        {
-            case IfStmtNode iff:
-                if (iff.Condition is not null && TryGetRange(iff.Condition, out var rc)) Acc(rc);
-                if (iff.Then is not null && TryGetRange(iff.Then, out var rt)) Acc(rt);
-                if (iff.Else is not null && TryGetRange(iff.Else, out var re)) Acc(re);
-                break;
-            case DoWhileStmtNode dw:
-                if (dw.Then is not null && TryGetRange(dw.Then, out var rdw)) Acc(rdw);
-                if (dw.Condition is not null && TryGetRange(dw.Condition, out var cdw)) Acc(cdw);
-                break;
-            case WhileStmtNode wl:
-                if (wl.Then is not null && TryGetRange(wl.Then, out var rwl)) Acc(rwl);
-                if (wl.Condition is not null && TryGetRange(wl.Condition, out var cwl)) Acc(cwl);
-                break;
-            case ForStmtNode fr:
-                if (fr.Init is not null && TryGetRange(fr.Init, out var ri)) Acc(ri);
-                if (fr.Condition is not null && TryGetRange(fr.Condition, out var rc2)) Acc(rc2);
-                if (fr.Increment is not null && TryGetRange(fr.Increment, out var rinc)) Acc(rinc);
-                if (fr.Then is not null && TryGetRange(fr.Then, out var rthen)) Acc(rthen);
-                break;
-            case ForeachStmtNode fe:
-                if (fe.Collection is not null && TryGetRange(fe.Collection, out var rcol)) Acc(rcol);
-                if (fe.Then is not null && TryGetRange(fe.Then, out var rfe)) Acc(rfe);
-                break;
-            case FunDevBlockNode fdb:
-                var rbody = GetStmtListRange(fdb.Body); Acc(rbody);
-                break;
-            case StmtListNode sl:
-                foreach (var st in sl.Statements)
-                {
-                    if (TryGetRange(st, out var rs)) Acc(rs);
-                }
-                break;
-            case SwitchStmtNode sw:
-                if (sw.Expression is not null && TryGetRange(sw.Expression, out var rexp)) Acc(rexp);
-                if (TryGetRange(sw.Cases, out var rcases)) Acc(rcases);
-                break;
-            case CaseListNode cl:
-                foreach (var cs in cl.Cases)
-                {
-                    if (TryGetRange(cs, out var rcs)) Acc(rcs);
-                }
-                break;
-            case CaseStmtNode cs:
-                foreach (var lbl in cs.Labels)
-                {
-                    if (TryGetRange(lbl, out var rlbl)) Acc(rlbl);
-                }
-                var rb = GetStmtListRange(cs.Body); Acc(rb);
-                break;
-            case CaseLabelNode cln when cln.Value is not null:
-                range = cln.Value.Range; return true;
-            case ClassBodyListNode cbl:
-                foreach (var d in cbl.Definitions)
-                {
-                    if (TryGetRange(d, out var rd)) Acc(rd);
-                }
-                break;
-        }
-
-        if (ok)
-        {
-            range = RangeHelper.From(start.Start, end.End);
-            return true;
-        }
-        range = default;
-        return false;
-    }
     public async Task AnalyseAsync(IEnumerable<IExportedSymbol> exportedSymbols, CancellationToken cancellationToken = default)
     {
         await WaitUntilParsedAsync(cancellationToken);
@@ -796,359 +402,6 @@ public class Script(DocumentUri ScriptUri, string languageId, ISymbolLocationPro
         }
     }
 
-    public async Task<Hover?> GetHoverAsync(Position position, CancellationToken cancellationToken)
-    {
-        if (!Sense.IsEditorMode) return null;
-        await WaitUntilAnalysedAsync(cancellationToken);
-
-        // Check if position is on a namespace qualifier (followed by :: and identifier)
-        // If so, forward to the function token for hover resolution
-        int initialIdx = Sense.Tokens.GetIndex(position);
-        Token? initialToken = Sense.Tokens.GetAt(initialIdx);
-        if (initialToken is not null && initialToken.Type == TokenType.Identifier)
-        {
-            if (TryGetQualifiedFunctionTokenByIndex(initialIdx, out int funcIdx))
-            {
-                // Forward position to the function token for hover lookup
-                position = Sense.Tokens.GetAt(funcIdx)!.Range.Start;
-            }
-        }
-
-        // Priority 0: Check HoverLibrary first for special cases (preprocessor macros, directives, etc.)
-        // These might not have regular tokens but still need hover support
-        IHoverable? hoverable = Sense.HoverLibrary!.Get(position);
-        if (hoverable is not null)
-        {
-            Log.Debug("[HOVER] Priority 0: Returning hoverable from library at position {Pos}", position);
-            return hoverable.GetHover();
-        }
-
-        Token? token = Sense.Tokens.Get(position);
-        if (token is null)
-        {
-            Log.Debug("[HOVER] No token found at position {Pos}", position);
-            return null;
-        }
-
-        Log.Debug("[HOVER] Found token: Type={Type}, Lexeme='{Lexeme}' at position {Pos}", 
-            token.Type, token.Lexeme, position);
-
-        // Priority 1: If inside a function call's parentheses, show signature with highlighted parameter
-        // But NOT if we're hovering on the function name itself - that should show the normal hover
-        if (TryGetCallInfo(token, out Token? callIdToken, out int activeParam))
-        {
-            // Check if we're hovering on the function identifier itself (not inside parentheses)
-            bool isOnFunctionIdentifier = token == callIdToken;
-
-            if (!isOnFunctionIdentifier)
-            {
-                Log.Debug("[HOVER] Priority 1: Inside function call, building signature hover for '{Func}' with active param {Param}", 
-                    callIdToken?.Lexeme ?? "unknown", activeParam);
-                return BuildCallSignatureHover(callIdToken, activeParam);
-            }
-            else
-            {
-                Log.Debug("[HOVER] Priority 1: On function identifier itself, skipping signature hover");
-            }
-        }
-
-        // Priority 2: If hovering on namespace qualifier, forward to the actual function
-        if (token.Type == TokenType.Identifier && TryGetQualifiedFunctionToken(token, out Token? functionToken))
-        {
-            Log.Debug("[HOVER] Priority 2: Namespace qualifier detected, forwarding to function token");
-            token = functionToken;
-        }
-
-        // Priority 3: Check HoverLibrary again for the resolved token
-        // (might be different from Priority 0 if we forwarded from namespace qualifier)
-        hoverable = Sense.HoverLibrary.Get(token.Range.Start);
-        if (hoverable is not null)
-        {
-            Log.Debug("[HOVER] Priority 3: Returning hoverable for resolved token at position {Pos}", token.Range.Start);
-            return hoverable.GetHover();
-        }
-
-        // Priority 4: Fallback - try to synthesize basic hover for script functions without SenseDefinition
-        // This handles edge cases where analysis might have missed something
-        if (token.Type == TokenType.Identifier && !IsBuiltinFunction(token.Lexeme))
-        {
-            return TryBuildFallbackFunctionHover(token);
-        }
-
-        return null;
-    }
-
-    /// <summary>
-    /// Checks if token is a namespace qualifier (e.g., "utility" in "utility::function")
-    /// and returns the actual function token after the ::
-    /// </summary>
-    private bool TryGetQualifiedFunctionToken(Token token, out Token? functionToken)
-    {
-        functionToken = null;
-
-        Token? next = token.Next;
-        while (next is not null && next.IsWhitespacey())
-        {
-            next = next.Next;
-        }
-
-        if (next is null || next.Type != TokenType.ScopeResolution)
-        {
-            return false;
-        }
-
-        Token? afterScope = next.Next;
-        while (afterScope is not null && afterScope.IsWhitespacey())
-        {
-            afterScope = afterScope.Next;
-        }
-
-        if (afterScope is not null && afterScope.Type == TokenType.Identifier)
-        {
-            functionToken = afterScope;
-            return true;
-        }
-
-        return false;
-    }
-
-    /// <summary>
-    /// Builds hover showing full function documentation when inside a call
-    /// </summary>
-    private Hover? BuildCallSignatureHover(Token functionToken, int activeParam)
-    {
-        var (qualifier, funcName) = ParseNamespaceQualifiedIdentifierByIndex(Sense.Tokens.IndexOf(functionToken));
-
-        Log.Debug("[HOVER] BuildCallSignatureHover: funcName='{FuncName}', qualifier='{Qualifier}', senseDef={SenseDefType}",
-            funcName, qualifier ?? "(none)", Sense.GetSenseDefinition(functionToken)?.GetType().Name ?? "null");
-
-        // Try to get the function definition to show its full documentation
-        ScrFunction? function = null;
-
-        // First, check if the function token has a SenseDefinition with the function
-        if (Sense.GetSenseDefinition(functionToken) is ScrFunctionSymbol funcSymbol)
-        {
-            Log.Debug("[HOVER] Found function via ScrFunctionSymbol");
-            function = funcSymbol.Source;
-        }
-        else if (Sense.GetSenseDefinition(functionToken) is ScrMethodSymbol methodSymbol)
-        {
-            Log.Debug("[HOVER] Found function via ScrMethodSymbol");
-            function = methodSymbol.Source;
-        }
-        else if (Sense.GetSenseDefinition(functionToken) is ScrFunctionReferenceSymbol refSymbol)
-        {
-            Log.Debug("[HOVER] Found ScrFunctionReferenceSymbol, looking up function in definitions table");
-            // Function reference doesn't contain the full function, need to look it up
-            string ns = qualifier ?? GetEffectiveNamespace();
-
-            // Try to get from local definitions first
-            var localFunc = DefinitionsTable?.LocalScopedFunctions
-                .FirstOrDefault(f => string.Equals(f.Function.Name, funcName, StringComparison.OrdinalIgnoreCase) &&
-                                   string.Equals(f.Function.Namespace, ns, StringComparison.OrdinalIgnoreCase));
-
-            if (localFunc != default)
-            {
-                Log.Debug("[HOVER] Found function in LocalScopedFunctions: ns={Ns}, name={Name}", ns, funcName);
-                function = localFunc.Value.Function;
-            }
-            else
-            {
-                // Try internal symbols (from other scripts)
-                string qualifiedName = string.IsNullOrEmpty(qualifier) ? funcName : $"{qualifier}::{funcName}";
-                if (DefinitionsTable?.InternalSymbols.TryGetValue(qualifiedName, out var symbol) == true && symbol is ScrFunction internalFunc)
-                {
-                    Log.Debug("[HOVER] Found function in InternalSymbols: qualifiedName={QualifiedName}", qualifiedName);
-                    function = internalFunc;
-                }
-            }
-        }
-
-        // If not in sense, check if it's a built-in API function
-        if (function is null)
-        {
-            Log.Debug("[HOVER] Trying API lookup for function '{FuncName}'", funcName);
-            var api = TryGetApi();
-            if (api is not null)
-            {
-                function = api.GetApiFunction(funcName);
-                if (function is not null)
-                {
-                    Log.Debug("[HOVER] Found function in API");
-                }
-            }
-        }
-
-        // If we found the function, use its full Documentation
-        if (function is not null)
-        {
-            Log.Information("[HOVER] Using function.Documentation for '{FuncName}'", funcName);
-            return new Hover
-            {
-                Range = functionToken.Range,
-                Contents = new MarkedStringsOrMarkupContent(new MarkupContent
-                {
-                    Kind = MarkupKind.Markdown,
-                    Value = function.Documentation
-                })
-            };
-        }
-
-        Log.Debug("[HOVER] Function not found, falling back to BuildSignatureMarkdown");
-
-        // Fallback: use simplified signature if function not found
-        string? markdown = BuildSignatureMarkdown(funcName, qualifier, activeParam);
-        if (markdown is null)
-            return null;
-
-        return new Hover
-        {
-            Range = functionToken.Range,
-            Contents = new MarkedStringsOrMarkupContent(new MarkupContent
-            {
-                Kind = MarkupKind.Markdown,
-                Value = markdown
-            })
-        };
-    }
-
-    /// <summary>
-    /// Last resort: build basic hover for script functions that don't have proper SenseDefinition
-    /// </summary>
-    private Hover? TryBuildFallbackFunctionHover(Token token)
-    {
-        int tokenIdx = Sense.Tokens.IndexOf(token);
-        var (qualifier, name) = tokenIdx >= 0
-            ? ParseNamespaceQualifiedIdentifierByIndex(tokenIdx)
-            : (null, token.Lexeme);
-
-        // Exclude builtin API functions
-        if (IsBuiltinFunction(name))
-        {
-            return null; // let existing hover (if any) handle API
-        }
-
-        // Find function/method in current script tables
-        string ns = qualifier ?? GetEffectiveNamespace();
-        string? doc = DefinitionsTable?.GetFunctionDoc(ns, name);
-        string[]? parameters = DefinitionsTable?.GetFunctionParameters(ns, name);
-
-        // If we don't even know it exists, give up
-        if (doc is null && parameters is null)
-        {
-            var anyLoc = DefinitionsTable?.GetFunctionLocationAnyNamespace(name);
-            if (anyLoc is null)
-            {
-                return null;
-            }
-            // Known to exist but no metadata - show minimal info
-            return new Hover
-            {
-                Range = token.Range,
-                Contents = new MarkedStringsOrMarkupContent(new MarkupContent
-                {
-                    Kind = MarkupKind.Markdown,
-                    Value = $"{s_gscCodeBlockStart}function {name}(){s_codeBlockEnd}"
-                })
-            };
-        }
-
-        // Build hover from available metadata
-        string[] cleanParams = parameters?.Select(StripDefault).ToArray() ?? Array.Empty<string>();
-        string signature = cleanParams.Length == 0
-            ? $"function {name}()"
-            : $"function {name}({string.Join(", ", cleanParams)})";
-
-        string content = string.IsNullOrWhiteSpace(doc)
-            ? $"{s_gscCodeBlockStart}{signature}{s_codeBlockEnd}"
-            : $"{s_gscCodeBlockStart}{signature}{s_codeBlockEnd}{s_markdownSeparator}{doc}";
-
-        return new Hover
-        {
-            Range = token.Range,
-            Contents = new MarkedStringsOrMarkupContent(new MarkupContent
-            {
-                Kind = MarkupKind.Markdown,
-                Value = content
-            })
-        };
-    }
-
-    private string? BuildSignatureMarkdown(string name, string? qualifier, int activeParam)
-    {
-        // Built-in API: try first
-        var api = TryGetApi();
-        if (api is not null)
-        {
-            try
-            {
-                var apiFn = api.GetApiFunction(name);
-                if (apiFn is not null)
-                {
-                    var overload = apiFn.Overloads.FirstOrDefault();
-                    var paramSeq = overload != null ? overload.Parameters : new List<ScrFunctionArg>();
-                    string[] names = paramSeq.Select(p => StripDefault(p.Name)).ToArray();
-                    string sig = FormatSignature(name, names, activeParam, qualifier);
-                    string desc = apiFn.Description ?? string.Empty;
-                    return string.IsNullOrEmpty(desc) ? sig : $"{sig}{s_markdownSeparator}{desc}";
-                }
-            }
-            catch { }
-        }
-
-        // Script-defined (local or imported)
-        string ns = qualifier ?? GetEffectiveNamespace();
-        string[]? parms = DefinitionsTable?.GetFunctionParameters(ns, name);
-        string? doc = DefinitionsTable?.GetFunctionDoc(ns, name);
-        if (parms is not null)
-        {
-            string sig = FormatSignature(name, parms.Select(StripDefault).ToArray(), activeParam, qualifier);
-            // Doc comment is already formatted by SanitizeDocForMarkdown, don't sanitize again
-            string formattedDoc = doc ?? string.Empty;
-            return string.IsNullOrEmpty(formattedDoc) ? sig : $"{sig}{s_markdownSeparator}{formattedDoc}";
-        }
-
-        // Fallback: show empty params signature if symbol exists somewhere
-        var any = DefinitionsTable?.GetFunctionLocationAnyNamespace(name);
-        if (any is not null)
-        {
-            return FormatSignature(name, Array.Empty<string>(), activeParam, qualifier);
-        }
-        return null;
-    }
-
-    private static string FormatSignature(string name, IReadOnlyList<string> parameters, int activeParam, string? qualifier)
-    {
-        string nsPrefix = string.IsNullOrEmpty(qualifier) ? string.Empty : qualifier + "::";
-        if (parameters.Count == 0)
-        {
-            return $"{s_gscCodeBlockStart}function {nsPrefix}{name}(){s_codeBlockEnd}";
-        }
-        var sb = new StringBuilder();
-        sb.Append(s_gscCodeBlockStart).Append("function ").Append(nsPrefix).Append(name).Append('(');
-        for (int i = 0; i < parameters.Count; i++)
-        {
-            if (i > 0) sb.Append(", ");
-            string p = StripDefault(parameters[i]);
-            if (i == activeParam)
-            {
-                sb.Append('<').Append(p).Append('>');
-            }
-            else
-            {
-                sb.Append(p);
-            }
-        }
-        sb.Append(')').Append(s_codeBlockEnd);
-        return sb.ToString();
-    }
-
-    private static bool TryGetCallInfo(Token token, out Token idToken, out int activeParam)
-    {
-        // Delegate to shared call context finder
-        return TryFindCallContext(token, out idToken, out activeParam) && idToken is not null;
-    }
-
     public async Task<CompletionList?> GetCompletionAsync(Position position, CancellationToken cancellationToken)
     {
         if (!Sense.IsEditorMode) return null;
@@ -1199,946 +452,5 @@ public class Script(DocumentUri ScriptUri, string languageId, ISymbolLocationPro
         var functions = DefinitionsTable.ExportedFunctions ?? [];
         var classes = DefinitionsTable.ExportedClasses ?? [];
         return functions.Cast<IExportedSymbol>().Concat(classes);
-    }
-
-    /// <summary>
-    /// Finds the call context from a given position in the token stream.
-    /// Scans backwards to find the opening '(' of a function call, identifies the function name,
-    /// and counts the active parameter index based on commas.
-    /// </summary>
-    /// <param name="token">Starting token (typically at cursor position)</param>
-    /// <param name="idToken">Output: the identifier token of the function being called</param>
-    /// <param name="activeParam">Output: the 0-based parameter index at the cursor position</param>
-    /// <returns>True if a valid call context was found, false otherwise</returns>
-    private static bool TryFindCallContext(Token token, out Token? idToken, out int activeParam)
-    {
-        idToken = null;
-        activeParam = 0;
-
-        // Scan left to find the nearest '(' that starts the current argument list
-        Token? cursor = token;
-        int parenDepth = 0;
-        while (cursor is not null)
-        {
-            if (cursor.Type == TokenType.CloseParen) parenDepth++;
-            if (cursor.Type == TokenType.OpenParen)
-            {
-                if (parenDepth == 0) break;
-                parenDepth--;
-            }
-            if (cursor.Type == TokenType.Identifier && cursor.Next?.Type == TokenType.OpenParen && parenDepth == 0)
-            {
-                cursor = cursor.Next;
-                break;
-            }
-            if (cursor.Type == TokenType.Semicolon || cursor.Type == TokenType.LineBreak)
-            {
-                return false; // Hit end of statement without finding '('
-            }
-            cursor = cursor.Previous;
-        }
-
-        if (cursor is null)
-            return false; // not in a call
-
-        // Find the identifier before this '('
-        Token? id = cursor.Previous;
-        while (id is not null && (id.IsWhitespacey() || id.IsComment())) 
-            id = id.Previous;
-
-        if (id is null || id.Type != TokenType.Identifier)
-            return false;
-
-        idToken = id;
-
-        // Count parameter index
-        // without nesting into inner parens/brackets/braces
-        Token? walker = cursor.Next;
-        int depthParen = 0, depthBracket = 0, depthBrace = 0;
-        int index = 0;
-        while (walker is not null && walker != token.Next)
-        {
-            if (walker.Type == TokenType.OpenParen) depthParen++;
-            else if (walker.Type == TokenType.CloseParen)
-            {
-                if (depthParen == 0) break; // end of this call
-                depthParen--;
-            }
-            else if (walker.Type == TokenType.OpenBracket) depthBracket++;
-            else if (walker.Type == TokenType.CloseBracket && depthBracket > 0) depthBracket--;
-            else if (walker.Type == TokenType.OpenBrace) depthBrace++;
-            else if (walker.Type == TokenType.CloseBrace && depthBrace > 0) depthBrace--;
-            else if (walker.Type == TokenType.Comma && depthParen == 0 && depthBracket == 0 && depthBrace == 0)
-            {
-                index++;
-            }
-            walker = walker.Next;
-        }
-
-        activeParam = index;
-        return true;
-    }
-
-    public async Task<Location?> GetDefinitionAsync(Position position, CancellationToken cancellationToken = default)
-    {
-        if (!Sense.IsEditorMode) return null;
-        await WaitUntilParsedAsync(cancellationToken);
-
-        Log.Debug("GetDefinitionAsync: Starting at position {Position}", position);
-
-        // First, allow preprocessor macro definitions/usages to resolve even if the original macro token was removed
-        IHoverable? hoverable = Sense.HoverLibrary!.Get(position);
-        if (hoverable is Pre.MacroDefinition macroDef && !macroDef.IsBuiltIn)
-        {
-            Log.Debug("GetDefinitionAsync: Found MacroDefinition at position");
-            string normalized = ScriptFileResolver.NormalizeFilePathForUri(ScriptUri.ToUri().LocalPath);
-            return new Location() { Uri = new Uri(normalized), Range = macroDef.Range };
-        }
-        if (hoverable is Pre.ScriptMacro scriptMacro)
-        {
-            Log.Debug("GetDefinitionAsync: Found ScriptMacro at position, navigating to definition");
-            string normalized = ScriptFileResolver.NormalizeFilePathForUri(ScriptUri.ToUri().LocalPath);
-            return new Location() { Uri = new Uri(normalized), Range = scriptMacro.DefineSource.Range };
-        }
-
-        int tokenIdx = Sense.Tokens.GetIndex(position);
-        Token? token = Sense.Tokens.GetAt(tokenIdx);
-        if (token is null)
-        {
-            Log.Debug("GetDefinitionAsync: No token found at position {Position}", position);
-            return null;
-        }
-
-        Log.Debug("GetDefinitionAsync: Found token '{Lexeme}' type={Type} at position {Position}, IsFromPreprocessor={IsPrep}", 
-            token.Lexeme, token.Type, position, token.IsFromPreprocessor);
-
-        // Log token range for debugging
-        Log.Debug("GetDefinitionAsync: Token range: Start=({Line}:{Char}), End=({EndLine}:{EndChar})", 
-            token.Range.Start.Line, token.Range.Start.Character, 
-            token.Range.End.Line, token.Range.End.Character);
-
-        // Log previous tokens for context
-        Token? prevToken = token.Previous;
-        int count = 0;
-        while (prevToken != null && count < 5)
-        {
-            Log.Debug("  Previous token [{Index}]: '{Lexeme}' type={Type}", count, prevToken.Lexeme, prevToken.Type);
-            prevToken = prevToken.Previous;
-            count++;
-        }
-
-        // Check if this is a variable reference - go to its definition
-        if (Sense.GetSenseDefinition(token) is ScrVariableSymbol varSymbol && varSymbol.DefinitionSource is not null)
-        {
-            Log.Debug("GetDefinitionAsync: Token is a variable reference, navigating to definition");
-            // Navigate to the definition (the identifier token where it was first declared)
-            string normalized = ScriptFileResolver.NormalizeFilePathForUri(ScriptUri.ToUri().LocalPath);
-            // Use the Range from the definition source's identifier token (stored in the variable symbol)
-            // We need to find the declaration token - for now, use the identifier token from definition source
-            Range targetRange = varSymbol.DefinitionSource switch
-            {
-                ExprNode expr => expr.Range,
-                ConstStmtNode constStmt => constStmt.IdentifierToken.Range,
-                ForeachStmtNode foreachStmt => varSymbol.IdentifierToken.Range, // Use the identifier token range
-                MemberDeclNode memberDecl => memberDecl.NameToken?.Range ?? varSymbol.IdentifierToken.Range,
-                ParamNode paramNode => paramNode.Name?.Range ?? varSymbol.IdentifierToken.Range, // Function parameter
-                _ => varSymbol.IdentifierToken.Range // Fallback to the identifier token
-            };
-            return new Location() { Uri = new Uri(normalized), Range = targetRange };
-        }
-
-        // Check if this is a parameter reference - go to its definition
-        if (Sense.GetSenseDefinition(token) is ScrParameterSymbol paramSymbol && paramSymbol.DefinitionSource is not null)
-        {
-            Log.Debug("GetDefinitionAsync: Token is a parameter reference, navigating to definition");
-            // Navigate to the parameter definition
-            string normalized = ScriptFileResolver.NormalizeFilePathForUri(ScriptUri.ToUri().LocalPath);
-            // For parameters, use the token from the ParamNode
-            Range targetRange = paramSymbol.DefinitionSource switch
-            {
-                ParamNode paramNode => paramNode.Name?.Range ?? paramSymbol.Range,
-                _ => paramSymbol.Range // Fallback to the parameter symbol's range
-            };
-            return new Location() { Uri = new Uri(normalized), Range = targetRange };
-        }
-
-        // If the token has an IntelliSense definition pointing at a dependency, return that file location.
-        if (Sense.GetSenseDefinition(token) is ScrDependencySymbol dep)
-        {
-            Log.Debug("GetDefinitionAsync: Token is a dependency symbol, navigating to file: {Path}", dep.Path);
-            string resolvedPath = dep.Path;
-            if (!File.Exists(resolvedPath))
-            {
-                Log.Debug("GetDefinitionAsync: Dependency file does not exist: {Path}", resolvedPath);
-                return null;
-            }
-            string normalized = ScriptFileResolver.NormalizeFilePathForUri(resolvedPath);
-            var targetUri = new Uri(normalized);
-            // Navigate to start of file in the target document
-            return new Location() { Uri = targetUri, Range = RangeHelper.From(0, 0, 0, 0) };
-        }
-
-        if (IsOnUsingLineByIndex(tokenIdx, out string? usingPath, out Range? usingRange))
-        {
-            string? resolved = Sense.GetDependencyPath(usingPath!, usingRange!);
-            if (resolved is not null && File.Exists(resolved))
-            {
-                string normalized = ScriptFileResolver.NormalizeFilePathForUri(resolved);
-                var targetUri = new Uri(normalized);
-                return new Location() { Uri = targetUri, Range = RangeHelper.From(0, 0, 0, 0) };
-            }
-        }
-
-        // If on an #insert line and we recorded a hover for it, go to the inserted file
-        IHoverable? h = Sense.HoverLibrary.Get(position);
-        if (h is Pre.InsertDirectiveHover ih)
-        {
-            string? resolved = Sense.ResolveInsertPath(ih.RawPath, ih.Range);
-            if (resolved is not null && File.Exists(resolved))
-            {
-                string normalized = ScriptFileResolver.NormalizeFilePathForUri(resolved);
-                var targetUri = new Uri(normalized);
-                return new Location() { Uri = targetUri, Range = RangeHelper.From(0, 0, 0, 0) };
-            }
-        }
-
-        // Ensure the token is a function-like identifier before attempting Go-to-Definition.
-        if (token.Type != TokenType.Identifier)
-        {
-            Log.Debug("GetDefinitionAsync: Token is not an identifier, type={Type}", token.Type);
-            return null;
-        }
-
-        Log.Debug("GetDefinitionAsync: Token is identifier '{Lexeme}', checking heuristics", token.Lexeme);
-
-        // If current token is namespace qualifier (followed by :: and identifier), forward to the function token
-        if (TryGetQualifiedFunctionTokenByIndex(tokenIdx, out int defFuncIdx))
-        {
-            tokenIdx = defFuncIdx;
-            token = Sense.Tokens.GetAt(tokenIdx)!;
-        }
-        int nextNonWsIdx = Sense.Tokens.NextNonWhitespaceIndex(tokenIdx);
-        Token? nextNonWs = Sense.Tokens.GetAt(nextNonWsIdx);
-        bool looksLikeCall = nextNonWs is not null && nextNonWs.Type == TokenType.OpenParen;
-        int prevNonTrivIdx = Sense.Tokens.PrevNonTriviaIndex(tokenIdx);
-        Token? prevNonTriv = Sense.Tokens.GetAt(prevNonTrivIdx);
-        bool isQualified = prevNonTriv is not null && prevNonTriv.Type == TokenType.ScopeResolution;
-        var tokenSenseDef = Sense.GetSenseDefinition(token);
-        bool hasDefinitionSymbol = tokenSenseDef is ScrFunctionSymbol || tokenSenseDef is ScrMethodSymbol || tokenSenseDef is ScrClassSymbol || tokenSenseDef is ScrClassReferenceSymbol;
-        bool isAddressOf = IsAddressOfIdentifierByIndex(tokenIdx);
-
-        Log.Debug("GetDefinitionAsync: Heuristics - looksLikeCall={Call}, isQualified={Qual}, hasDefinitionSymbol={Def}, isAddressOf={Addr}", 
-            looksLikeCall, isQualified, hasDefinitionSymbol, isAddressOf);
-
-        if (!looksLikeCall && !isQualified && !hasDefinitionSymbol && !isAddressOf)
-        {
-            Log.Debug("GetDefinitionAsync: Failed heuristic checks, returning null");
-            return null;
-        }
-
-        var (qualifier, name) = ParseNamespaceQualifiedIdentifierByIndex(tokenIdx);
-        Log.Debug("GetDefinitionAsync: Parsed identifier - qualifier={Qualifier}, name={Name}", qualifier ?? "(none)", name);
-
-        if (IsBuiltinFunction(name))
-        {
-            Log.Debug("GetDefinitionAsync: Identifier is builtin function, returning null");
-            return null;
-        }
-        if (qualifier is not null && DefinitionsTable is not null)
-        {
-            Log.Debug("GetDefinitionAsync: Looking up qualified function/class: {Qualifier}::{Name}", qualifier, name);
-            var loc = DefinitionsTable.GetFunctionLocation(qualifier, name)
-                   ?? DefinitionsTable.GetClassLocation(qualifier, name);
-            if (loc is not null)
-            {
-                Log.Debug("GetDefinitionAsync: Found qualified definition at {File}:{Range}", loc.Value.FilePath, loc.Value.Range);
-                string currentScriptPath = ScriptUri.ToUri().LocalPath;
-                return ScriptFileResolver.ResolveDefinitionLocation(currentScriptPath, loc.Value.FilePath, loc.Value.Range.ToRange());
-            }
-            Log.Debug("GetDefinitionAsync: Qualified lookup failed");
-        }
-        string ns = GetEffectiveNamespace();
-        Log.Debug("GetDefinitionAsync: Looking up in current namespace: {Namespace}", ns);
-        var localLoc = DefinitionsTable?.GetFunctionLocation(ns, name)
-                    ?? DefinitionsTable?.GetClassLocation(ns, name);
-        if (localLoc is not null)
-        {
-            Log.Debug("GetDefinitionAsync: Found local definition at {File}:{Range}", localLoc.Value.FilePath, localLoc.Value.Range);
-            string currentScriptPath = ScriptUri.ToUri().LocalPath;
-            return ScriptFileResolver.ResolveDefinitionLocation(currentScriptPath, localLoc.Value.FilePath, localLoc.Value.Range.ToRange());
-        }
-        Log.Debug("GetDefinitionAsync: Local namespace lookup failed, trying any namespace");
-        var anyLoc = DefinitionsTable?.GetFunctionLocationAnyNamespace(name)
-                  ?? DefinitionsTable?.GetClassLocationAnyNamespace(name);
-        if (anyLoc is not null)
-        {
-            Log.Debug("GetDefinitionAsync: Found definition in any namespace at {File}:{Range}", anyLoc.Value.FilePath, anyLoc.Value.Range);
-            string currentScriptPath = ScriptUri.ToUri().LocalPath;
-            return ScriptFileResolver.ResolveDefinitionLocation(currentScriptPath, anyLoc.Value.FilePath, anyLoc.Value.Range.ToRange());
-        }
-        Log.Debug("GetDefinitionAsync: All lookups failed, returning null");
-        return null;
-    }
-
-    public async Task<(string? qualifier, string name)?> GetQualifiedIdentifierAtAsync(Position position, CancellationToken cancellationToken = default)
-    {
-        if (!Sense.IsEditorMode) return null;
-        await WaitUntilParsedAsync(cancellationToken);
-
-        int idx = Sense.Tokens.GetIndex(position);
-        Token? token = Sense.Tokens.GetAt(idx);
-        if (token is null)
-        {
-            return null;
-        }
-
-        return ParseNamespaceQualifiedIdentifierByIndex(idx);
-    }
-
-    private static string? ExtractParameterDocFromDoc(string? doc, string paramName, int paramIndex)
-    {
-        if (string.IsNullOrWhiteSpace(doc) || string.IsNullOrWhiteSpace(paramName)) return null;
-        string Normalize(string s) => s.Trim().Trim('<', '>', '[', ']').Trim();
-
-        // Try to parse prototype line to map by position
-        string[] ParseDocPrototypeParams(string d)
-        {
-            var lines = d.Replace("\r\n", "\n").Replace('\r', '\n').Split('\n');
-            var line = lines.Length > 1 ? lines[1].Trim() : string.Empty; // second line from ```gsc\r\nfoo(bar)...```
-            int lp = line.IndexOf('(');
-            int rp = line.LastIndexOf(')');
-            if (lp < 0 || rp < lp) return Array.Empty<string>();
-            string inside = line.Substring(lp + 1, rp - lp - 1);
-            if (string.IsNullOrWhiteSpace(inside)) return Array.Empty<string>();
-            var parts = inside.Split(',');
-            var list = new List<string>(parts.Length);
-            foreach (var p in parts)
-            {
-                var v = Normalize(p);
-                if (v.Length > 0) list.Add(v);
-            }
-            return list.ToArray();
-        }
-
-        string[] protoParams = ParseDocPrototypeParams(doc);
-        List<string> candidates = new() { Normalize(paramName) };
-        if (paramIndex >= 0 && paramIndex < protoParams.Length)
-        {
-            candidates.Add(Normalize(protoParams[paramIndex]));
-        }
-
-        // Scan parameter description lines (``<param>`` — desc)
-        string[] linesDoc = doc.Replace("\r\n", "\n").Replace('\r', '\n').Split('\n');
-        foreach (var raw in linesDoc)
-        {
-            string line = raw.Trim();
-            if (line.Length == 0) continue;
-            int b1 = line.IndexOf('`');
-            if (b1 < 0) continue;
-            int b2 = line.IndexOf('`', b1 + 1);
-            if (b2 < 0) continue;
-            string token = Normalize(line.Substring(b1 + 1, b2 - b1 - 1));
-            bool match = false;
-            foreach (var c in candidates)
-            {
-                if (string.Equals(c, token, StringComparison.OrdinalIgnoreCase)) { match = true; break; }
-            }
-            if (!match) continue;
-            int dash = line.IndexOf('—', b2 + 1); // em dash
-            if (dash < 0) dash = line.IndexOf('-', b2 + 1); // fallback
-            string desc = dash >= 0 && dash + 1 < line.Length ? line[(dash + 1)..].Trim() : string.Empty;
-            return desc;
-        }
-        return null;
-    }
-
-    private bool TryGetCallInfoByIndex(int tokenIdx, out int idTokenIdx, out int activeParam)
-    {
-        idTokenIdx = -1;
-        activeParam = 0;
-        Token? token = Sense.Tokens.GetAt(tokenIdx);
-        if (token is null) return false;
-        if (!TryFindCallContext(token, out Token? idToken, out activeParam)) return false;
-        if (idToken is null) return false;
-        idTokenIdx = Sense.Tokens.IndexOf(idToken);
-        return idTokenIdx >= 0;
-    }
-
-    public async Task<SignatureHelp?> GetSignatureHelpAsync(Position position, CancellationToken cancellationToken)
-    {
-        await WaitUntilParsedAsync(cancellationToken);
-
-        var tokens = Sense.Tokens;
-        int tokenIdx = tokens.GetIndex(position);
-        Token? token = tokens.GetAt(tokenIdx);
-        if (token is null)
-            return null;
-
-        // Use index-based call info detection
-        if (!TryGetCallInfoByIndex(tokenIdx, out int idIdx, out int activeParam))
-            return null;
-
-        Token? id = tokens.GetAt(idIdx);
-        if (id is null)
-            return null;
-
-        var (qualifier, name) = ParseNamespaceQualifiedIdentifierByIndex(idIdx);
-
-        // Try builtin API first
-        List<SignatureInformation> signatures = new();
-        var api = TryGetApi();
-        if (api is not null)
-        {
-            try
-            {
-                var apiFn = api.GetApiFunction(name);
-                if (apiFn is not null)
-                {
-                    var docContent = new MarkupContent { Kind = MarkupKind.Markdown, Value = apiFn.Description ?? string.Empty };
-                    foreach (var overload in apiFn.Overloads)
-                    {
-                        IEnumerable<ScrFunctionArg> paramSeq = overload.Parameters;
-                        string calledOnStr = overload.CalledOn is ScrFunctionArg co ? $"{co.Name} " : string.Empty;
-                        var cleaned = paramSeq.Select(p => StripDefault(p.Name)).ToArray();
-                        string label = $"{calledOnStr}function {name}({string.Join(", ", cleaned)})";
-                        var parameters = new Container<ParameterInformation>(paramSeq.Select(p => new ParameterInformation { Label = StripDefault(p.Name), Documentation = string.IsNullOrWhiteSpace(p.Description) ? null : new MarkupContent { Kind = MarkupKind.Markdown, Value = p.Description! } }));
-                        signatures.Add(new SignatureInformation { Label = label, Documentation = docContent, Parameters = parameters });
-                    }
-                }
-            }
-            catch { }
-        }
-
-        // Then script-defined (local or imported) using DefinitionsTable
-        string ns = qualifier ?? GetEffectiveNamespace();
-        string[]? parms = DefinitionsTable?.GetFunctionParameters(ns, name) ?? DefinitionsTable?.GetFunctionParameters(qualifier ?? ns, name);
-        string? doc = DefinitionsTable?.GetFunctionDoc(ns, name);
-        if (parms is not null)
-        {
-            var cleaned = parms.Select(StripDefault).ToArray();
-            string label = $"function {name}({string.Join(", ", cleaned)})";
-            var paramList = new List<ParameterInformation>(cleaned.Length);
-            for (int i = 0; i < cleaned.Length; i++)
-            {
-                string p = cleaned[i];
-                string? pDoc = ExtractParameterDocFromDoc(doc, p, i);
-                paramList.Add(new ParameterInformation
-                {
-                    Label = p,
-                    Documentation = string.IsNullOrWhiteSpace(pDoc) ? null : new MarkupContent { Kind = MarkupKind.Markdown, Value = pDoc }
-                });
-            }
-            var parameters = new Container<ParameterInformation>(paramList);
-            // Do not include full Markdown doc in SignatureHelp for script-defined; keep prototype and parameters only
-            signatures.Add(new SignatureInformation { Label = label, Parameters = parameters });
-        }
-
-        if (signatures.Count == 0)
-            return null;
-
-        // Find the best matching signature based on parameter count
-        int activeSignature = 0;
-        for (int i = 0; i < signatures.Count; i++)
-        {
-            if (signatures[i].Parameters is { } p && p.Count() > activeParam)
-            {
-                activeSignature = i;
-                break;
-            }
-        }
-
-        int paramCount = 1;
-        if (signatures[activeSignature].Parameters is { } paramContainer)
-        {
-            paramCount = paramContainer.Count();
-        }
-
-        return new SignatureHelp
-        {
-            ActiveSignature = activeSignature,
-            ActiveParameter = Math.Max(0, Math.Min(activeParam, paramCount - 1)),
-            Signatures = new Container<SignatureInformation>(signatures)
-        };
-    }
-
-    private static string StripDefault(string? name)
-    {
-        if (string.IsNullOrWhiteSpace(name)) return string.Empty;
-        int idx = name.IndexOf('=');
-        return idx >= 0 ? name[..idx].Trim() : name.Trim();
-    }
-
-    // ===== SPA helpers =====
-
-    private void EmitUnusedParameterDiagnostics()
-    {
-        if (RootNode is null) return;
-        // Traverse functions
-        foreach (var fn in EnumerateFunctions(RootNode))
-        {
-            // collect used identifiers in function body
-            HashSet<string> used = new(StringComparer.OrdinalIgnoreCase);
-            CollectIdentifiers(fn.Body, used);
-            foreach (var p in fn.Parameters.Parameters)
-            {
-                if (p.Name is null) continue;
-                string paramName = p.Name.Lexeme;
-                if (!used.Contains(paramName))
-                {
-                    Sense.AddSpaDiagnostic(p.Name.Range, GSCErrorCodes.UnusedParameter, paramName);
-                }
-            }
-        }
-    }
-
-    private void EmitUnusedUsingDiagnostics()
-    {
-        if (RootNode is null || DefinitionsTable is null) return;
-
-        // Map referenced file paths
-        HashSet<string> referencedFiles = new(StringComparer.OrdinalIgnoreCase);
-        foreach (var kv in _references)
-        {
-            var key = kv.Key;
-            var fLoc = DefinitionsTable.GetFunctionLocation(key.Namespace, key.Name);
-            if (fLoc is not null) { referencedFiles.Add(ScriptFileResolver.NormalizeFilePathForUri(fLoc.Value.FilePath)); continue; }
-            var cLoc = DefinitionsTable.GetClassLocation(key.Namespace, key.Name);
-            if (cLoc is not null) { referencedFiles.Add(ScriptFileResolver.NormalizeFilePathForUri(cLoc.Value.FilePath)); continue; }
-        }
-
-        // For each using directive, if no referenced symbol comes from that dependency file, mark unused
-        foreach (var depNode in RootNode.Dependencies)
-        {
-            // Build expected suffix: ..\scripts\<depNode.Path>.<LanguageId>
-            string rel = depNode.Path.Replace('/', Path.DirectorySeparatorChar).Replace('\\', Path.DirectorySeparatorChar).TrimStart(Path.DirectorySeparatorChar);
-            string expectedSuffix = Path.DirectorySeparatorChar + rel + "." + LanguageId;
-
-            bool anyFromThisUsing = false;
-            foreach (var referenced in referencedFiles)
-            {
-                // referenced already normalized
-                if (referenced.Contains(expectedSuffix, StringComparison.OrdinalIgnoreCase))
-                {
-                    anyFromThisUsing = true;
-                    break;
-                }
-            }
-
-            if (!anyFromThisUsing)
-            {
-                Sense.AddSpaDiagnostic(depNode.Range, GSCErrorCodes.UnusedUsing, depNode.Path);
-            }
-        }
-    }
-
-    private void EmitUnusedVariableDiagnostics()
-    {
-        if (RootNode is null) return;
-        foreach (var fn in EnumerateFunctions(RootNode))
-        {
-            // Count all identifier occurrences within the function body
-            Dictionary<string, int> usageCounts = new(StringComparer.OrdinalIgnoreCase);
-            CollectIdentifierCounts(fn.Body, usageCounts);
-
-            // Single pass through function body to flag unused consts and assignments
-            foreach (var node in EnumerateChildren(fn.Body))
-            {
-                if (node is ConstStmtNode cst)
-                {
-                    string name = cst.Identifier;
-                    usageCounts.TryGetValue(name, out int count);
-                    if (count == 0)
-                    {
-                        Sense.AddSpaDiagnostic(cst.Range, GSCErrorCodes.UnusedVariable, name);
-                    }
-                    continue;
-                }
-
-                if (node is ExprStmtNode es && es.Expr is BinaryExprNode be && be.Operation == TokenType.Assign && be.Left is IdentifierExprNode id)
-                {
-                    string name = id.Identifier;
-                    usageCounts.TryGetValue(name, out int count);
-                    // If the only occurrence is this defining assignment (LHS), consider it unused
-                    if (count <= 1)
-                    {
-                        Sense.AddSpaDiagnostic(id.Range, GSCErrorCodes.UnusedVariable, name);
-                    }
-                }
-            }
-        }
-    }
-
-    private static void CollectIdentifierCounts(AstNode node, Dictionary<string, int> into)
-    {
-        if (node is IdentifierExprNode id)
-        {
-            if (!into.TryGetValue(id.Identifier, out int c)) c = 0;
-            into[id.Identifier] = c + 1;
-        }
-        foreach (var child in EnumerateChildren(node))
-        {
-            CollectIdentifierCounts(child, into);
-        }
-    }
-
-    private void EmitSwitchCaseDiagnostics()
-    {
-        // Switch case diagnostics (duplicate labels, fallthrough, multiple defaults)
-        // are now handled in ControlFlowAnalyser (CFA).
-        // This method is kept as a placeholder for any future switch-specific diagnostics
-        // that don't fit in CFA.
-    }
-
-    private static IEnumerable<SwitchStmtNode> EnumerateSwitches(AstNode node)
-    {
-        if (node is SwitchStmtNode s) yield return s;
-        foreach (var child in EnumerateChildren(node))
-        {
-            foreach (var x in EnumerateSwitches(child)) yield return x;
-        }
-    }
-
-    // Enumerators
-    private static IEnumerable<FunDefnNode> EnumerateFunctions(AstNode node)
-    {
-        if (node is FunDefnNode fn) yield return fn;
-        foreach (var child in EnumerateChildren(node))
-        {
-            foreach (var f in EnumerateFunctions(child)) yield return f;
-        }
-    }
-
-    private static IEnumerable<FunCallNode> EnumerateCalls(AstNode node)
-    {
-        if (node is FunCallNode f) yield return f;
-        foreach (var child in EnumerateChildren(node))
-        {
-            foreach (var c in EnumerateCalls(child)) yield return c;
-        }
-    }
-
-    private static IEnumerable<NamespacedMemberNode> EnumerateNamespacedMembers(AstNode node)
-    {
-        if (node is NamespacedMemberNode n) yield return n;
-        foreach (var child in EnumerateChildren(node))
-        {
-            foreach (var c in EnumerateNamespacedMembers(child)) yield return c;
-        }
-    }
-
-    private static IEnumerable<BinaryExprNode> EnumerateBinaryExprs(AstNode node)
-    {
-        if (node is BinaryExprNode b) yield return b;
-        foreach (var child in EnumerateChildren(node))
-        {
-            foreach (var c in EnumerateBinaryExprs(child)) yield return c;
-        }
-    }
-
-    private static bool ContainsThreadCall(AstNode node)
-    {
-        if (node is PrefixExprNode pe && pe.Operation == TokenType.Thread)
-        {
-            return true;
-        }
-        if (node is CalledOnNode con)
-        {
-            // thread appears on the call side for patterns like self thread foo();
-            if (ContainsThreadCall(con.Call)) return true;
-            // also traverse 'on' to be thorough
-            if (ContainsThreadCall(con.On)) return true;
-            return false;
-        }
-        foreach (var child in EnumerateChildren(node))
-        {
-            if (ContainsThreadCall(child)) return true;
-        }
-        return false;
-    }
-
-    private static bool ContainsThreadCallCached(AstNode node, Dictionary<AstNode, bool> cache)
-    {
-        if (cache.TryGetValue(node, out var v)) return v;
-        bool result;
-        if (node is PrefixExprNode pe && pe.Operation == TokenType.Thread)
-        {
-            result = true;
-        }
-        else if (node is CalledOnNode con)
-        {
-            result = ContainsThreadCallCached(con.Call, cache) || ContainsThreadCallCached(con.On, cache);
-        }
-        else
-        {
-            result = false;
-            foreach (var child in EnumerateChildren(node))
-            {
-                if (ContainsThreadCallCached(child, cache)) { result = true; break; }
-            }
-        }
-        cache[node] = result;
-        return result;
-    }
-
-    private void EmitAssignOnThreadDiagnostics()
-    {
-        if (RootNode is null) return;
-
-        // Early exit: Check if there are any thread calls in the entire file first
-        // This avoids expensive nested enumeration if there's nothing to check
-        bool hasAnyThreadCalls = false;
-        foreach (var node in EnumerateChildren(RootNode))
-        {
-            if (ContainsThreadCallQuickCheck(node))
-            {
-                hasAnyThreadCalls = true;
-                break;
-            }
-        }
-
-        if (!hasAnyThreadCalls)
-        {
-            return; // No thread calls in file, skip expensive analysis
-        }
-
-        // Proceed with full analysis only if thread calls exist
-        var cache = new Dictionary<AstNode, bool>(ReferenceEqualityComparer.Instance);
-        foreach (var fn in EnumerateFunctions(RootNode))
-        {
-            foreach (var bin in EnumerateBinaryExprs(fn.Body))
-            {
-                if (bin.Operation == TokenType.Assign && bin.Right is not null)
-                {
-                    if (ContainsThreadCallCached(bin.Right, cache))
-                    {
-                        Sense.AddSpaDiagnostic(bin.Range, GSCErrorCodes.AssignOnThreadedFunction);
-                    }
-                }
-            }
-        }
-    }
-
-    // Quick shallow check for thread token without deep recursion
-    private static bool ContainsThreadCallQuickCheck(AstNode node)
-    {
-        if (node is PrefixExprNode pe && pe.Operation == TokenType.Thread)
-        {
-            return true;
-        }
-        // Only check immediate children, not deep recursion
-        foreach (var child in EnumerateChildren(node))
-        {
-            if (child is PrefixExprNode pec && pec.Operation == TokenType.Thread)
-            {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private static IEnumerable<AstNode> EnumerateChildren(AstNode node)
-    {
-        switch (node)
-        {
-            case ScriptNode s:
-                foreach (var d in s.Dependencies) yield return d;
-                foreach (var sd in s.ScriptDefns) yield return sd; break;
-            case ClassBodyListNode cbl:
-                foreach (var d in cbl.Definitions) yield return d; break;
-            case StmtListNode sl:
-                foreach (var st in sl.Statements) yield return st; break;
-            case IfStmtNode iff:
-                if (iff.Condition is not null) yield return iff.Condition;
-                if (iff.Then is not null) yield return iff.Then;
-                if (iff.Else is not null) yield return iff.Else; break;
-            case ReservedFuncStmtNode rfs:
-                if (rfs.Expr is not null) yield return rfs.Expr; break;
-            case ConstStmtNode cst:
-                if (cst.Value is not null) yield return cst.Value; break;
-            case ExprStmtNode es:
-                if (es.Expr is not null) yield return es.Expr; break;
-            case DoWhileStmtNode dw:
-                if (dw.Condition is not null) yield return dw.Condition; if (dw.Then is not null) yield return dw.Then; break;
-            case WhileStmtNode wl:
-                if (wl.Condition is not null) yield return wl.Condition; if (wl.Then is not null) yield return wl.Then; break;
-            case ForStmtNode fr:
-                if (fr.Init is not null) yield return fr.Init; if (fr.Condition is not null) yield return fr.Condition; if (fr.Increment is not null) yield return fr.Increment; if (fr.Then is not null) yield return fr.Then; break;
-            case ForeachStmtNode fe:
-                if (fe.Collection is not null) yield return fe.Collection; if (fe.Then is not null) yield return fe.Then; break;
-            case ReturnStmtNode rt:
-                if (rt.Value is not null) yield return rt.Value; break;
-            case DefnDevBlockNode db:
-                foreach (var d in db.Definitions) yield return d; break;
-            case FunDevBlockNode fdb:
-                yield return fdb.Body; break;
-            case SwitchStmtNode sw:
-                if (sw.Expression is not null) yield return sw.Expression; yield return sw.Cases; break;
-            case CaseListNode cl:
-                foreach (var c in cl.Cases) yield return c; break;
-            case CaseStmtNode cs:
-                foreach (var l in cs.Labels) yield return l; yield return cs.Body; break;
-            case CaseLabelNode cln:
-                if (cln.Value is not null) yield return cln.Value; break;
-            case FunDefnNode fd:
-                yield return fd.Parameters; yield return fd.Body; break;
-            case ParamListNode pl:
-                foreach (var p in pl.Parameters) yield return p; break;
-            case ParamNode pn:
-                if (pn.Default is not null) yield return pn.Default; break;
-            case VectorExprNode vx:
-                yield return vx.X; if (vx.Y is not null) yield return vx.Y; if (vx.Z is not null) yield return vx.Z; break;
-            case TernaryExprNode te:
-                yield return te.Condition; if (te.Then is not null) yield return te.Then; if (te.Else is not null) yield return te.Else; break;
-            case BinaryExprNode be:
-                if (be.Left is not null) yield return be.Left; if (be.Right is not null) yield return be.Right; break;
-            case PrefixExprNode pe:
-                yield return pe.Operand; break;
-            case PostfixExprNode pxe:
-                yield return pxe.Operand; break;
-            case ConstructorExprNode:
-                break;
-            case MethodCallNode mc:
-                if (mc.Target is not null) yield return mc.Target; yield return mc.Arguments; break;
-            case FunCallNode fc:
-                if (fc.Function is not null) yield return fc.Function; yield return fc.Arguments; break;
-            case NamespacedMemberNode nm:
-                yield return nm.Namespace; yield return nm.Member; break;
-            case ArgsListNode al:
-                foreach (var a in al.Arguments) { if (a is not null) yield return a; }
-                break;
-            case ArrayIndexNode ai:
-                yield return ai.Array; if (ai.Index is not null) yield return ai.Index; break;
-            case CalledOnNode con:
-                yield return con.On; yield return con.Call; break;
-            default:
-                break;
-        }
-    }
-
-    private static void CollectIdentifiers(AstNode node, HashSet<string> into)
-    {
-        switch (node)
-        {
-            case IdentifierExprNode id:
-                into.Add(id.Identifier);
-                break;
-        }
-        foreach (var child in EnumerateChildren(node))
-        {
-            CollectIdentifiers(child, into);
-        }
-    }
-
-    public async Task<IReadOnlyList<Range>> GetLocalVariableReferencesAsync(Position position, bool includeDeclaration, CancellationToken cancellationToken = default)
-    {
-        if (!Sense.IsEditorMode) return Array.Empty<Range>();
-        await WaitUntilParsedAsync(cancellationToken);
-
-        // Acquire the token under the cursor
-        Token? token = Sense.Tokens.Get(position);
-        if (token is null || token.Type != TokenType.Identifier)
-        {
-            return Array.Empty<Range>();
-        }
-        string target = token.Lexeme;
-
-        // Find the enclosing function scope
-        FunctionScopeInfo? enclosingScope = null;
-
-        if (_functionScopes is not null)
-        {
-            // Post-analysis: use precomputed scopes
-            foreach (var scope in _functionScopes)
-            {
-                if (IsPositionInsideRange(position, scope.BodyRange))
-                {
-                    enclosingScope = scope;
-                    break;
-                }
-                foreach (var (pName, pRange) in scope.Parameters)
-                {
-                    if (ComparePosition(position, pRange.Start) >= 0 && ComparePosition(pRange.End, position) >= 0)
-                    {
-                        enclosingScope = scope;
-                        break;
-                    }
-                }
-                if (enclosingScope is not null) break;
-            }
-        }
-        else if (RootNode is not null)
-        {
-            // Pre-analysis fallback: compute from AST
-            foreach (var fn in EnumerateFunctions(RootNode))
-            {
-                var bodyRange = GetStmtListRange(fn.Body);
-                if (IsPositionInsideRange(position, bodyRange))
-                {
-                    var parms = new List<(string Name, Range Range)>();
-                    foreach (var p in fn.Parameters.Parameters)
-                    {
-                        if (p.Name is not null) parms.Add((p.Name.Lexeme, p.Name.Range));
-                    }
-                    enclosingScope = new FunctionScopeInfo(fn.Name?.Lexeme, bodyRange, parms);
-                    break;
-                }
-                foreach (var p in fn.Parameters.Parameters)
-                {
-                    if (p.Name is null) continue;
-                    if (ComparePosition(position, p.Name.Range.Start) >= 0 && ComparePosition(p.Name.Range.End, position) >= 0)
-                    {
-                        var parms = new List<(string Name, Range Range)>();
-                        foreach (var pp in fn.Parameters.Parameters)
-                        {
-                            if (pp.Name is not null) parms.Add((pp.Name.Lexeme, pp.Name.Range));
-                        }
-                        enclosingScope = new FunctionScopeInfo(fn.Name?.Lexeme, GetStmtListRange(fn.Body), parms);
-                        break;
-                    }
-                }
-                if (enclosingScope is not null) break;
-            }
-        }
-
-        if (enclosingScope is null)
-        {
-            return Array.Empty<Range>();
-        }
-
-        // Collect identifier tokens within function body matching the name
-        Range body = enclosingScope.BodyRange;
-        var results = new List<Range>();
-        var tokensLib = Sense.Tokens;
-        for (int i = 0; i < tokensLib.Count; i++)
-        {
-            Token t = tokensLib.GetAt(i)!;
-            if (t.Type != TokenType.Identifier) continue;
-            // Restrict to tokens in the same function body
-            if (!(ComparePosition(t.Range.Start, body.Start) >= 0 && ComparePosition(body.End, t.Range.End) >= 0))
-                continue;
-            // Match name (case-insensitive to be consistent with SPA checks)
-            if (!string.Equals(t.Lexeme, target, StringComparison.OrdinalIgnoreCase)) continue;
-            // Exclude namespace-qualified identifiers or function calls
-            int prevIdx = tokensLib.PrevNonTriviaIndex(i);
-            if (prevIdx >= 0 && tokensLib.GetAt(prevIdx)!.Type == TokenType.ScopeResolution) continue;
-            int nextIdx = tokensLib.NextNonWhitespaceIndex(i);
-            if (nextIdx >= 0 && tokensLib.GetAt(nextIdx)!.Type == TokenType.OpenParen) continue;
-
-            results.Add(t.Range);
-        }
-
-        // Optionally include declaration site (parameter declaration)
-        if (includeDeclaration)
-        {
-            foreach (var (pName, pRange) in enclosingScope.Parameters)
-            {
-                if (string.Equals(pName, target, StringComparison.OrdinalIgnoreCase))
-                {
-                    results.Add(pRange);
-                }
-            }
-        }
-
-        return results;
     }
 }
